Incfrease the input limit to long

+++ problem_solving/ChessBoardSquaresCount.java
@@ -6,20 +6,21 @@ import java.io.*;
 
 class ChessBoardSquaresCount {
 
-	private static int nC2(int n) {
+	private static long nC2(long n) {
 		return (n * (n-1)) / 2;
 	}
 
 	// Take two horizontal lines and two vertical lines 
 	// for the creation of rectangle.
 	// n+1 - because there are n+1 lines needed to create n sized rectangles.
-	private static int rectanglesCount(int n) {
+	// Includes the squares as well since a square is also a rectangle.
+	private static long rectanglesCount(long n) {
 		return nC2(n+1) * nC2(n+1);
 	}
 
-    private static int squaresCount(int n) {
-        int result = 0;
-        for (int i = 0; i < n; i++) {
+    private static long squaresCount(long n) {
+        long result = 0;
+        for (long i = 0; i < n; i++) {
             result += (n-i) * (n-i);
         }
         return result;
@@ -31,7 +32,7 @@ class ChessBoardSquaresCount {
 		int t = sc.nextInt();
 		while (t > 0) {
 		    t -= 1;
-		    int n = sc.nextInt();
+		    long n = sc.nextLong();
 		    System.out.println(squaresCount(n));
 		    System.out.println(rectanglesCount(n));
 		}
